Extract WTA CLI handlers
Move probe, hooks, and sessions command execution into focused CLI modules while keeping top-level dispatch in main.rs.

Closes #523


Copilot-Session: 725545ce-7285-44f9-ab47-538b3149d444
