+++ tools/wta/src/cli/hooks.rs
@@ -1,0 +1,157 @@
+use anyhow::Result;
+
+use super::args::HooksCliFilter;
+
+pub(crate) fn run_install(cli: HooksCliFilter) -> Result<()> {
+    // Logging is initialized in `main()`; the install attempt is observable in
+    // %LOCALAPPDATA%\IntelligentTerminal\logs\wta-install-hooks.log.
+    let scope = cli.into_scope();
+    crate::agent_hooks_installer::ensure_installed_scoped(scope);
+
+    // Verify the install actually landed by checking on-disk status.
+    // ensure_installed_scoped is fire-and-forget (silent on failure),
+    // so we inspect the result independently. `status_scoped(scope)`
+    // skips the Node-CLI spawns for CLIs outside the requested scope —
+    // a `--cli copilot` install no longer pays for `claude plugin list`
+    // and `gemini extensions list` (each ~1-3s of Node startup).
+    let report = crate::agent_hooks_installer::status_scoped(scope);
+    let failed: Vec<&str> = report
+        .clis
+        .iter()
+        .filter(|c| {
+            let in_scope = match scope {
+                crate::agent_hooks_installer::CliScope::All => true,
+                crate::agent_hooks_installer::CliScope::One(kind) => c.name == kind.name(),
+            };
+            // A CLI is "failed" if it's in scope, present on the machine
+            // (cli_found), but hooks are not installed.
+            in_scope && c.binary_on_path && !c.plugin_installed
+        })
+        .map(|c| c.name)
+        .collect();
+
+    if failed.is_empty() {
+        println!("{}", t!("hooks.install_attempted"));
+        Ok(())
+    } else {
+        let names = failed.join(", ");
+        tracing::error!(target: "agent_hooks", clis = %names, "hooks install verification failed");
+        anyhow::bail!("hooks installation failed for: {}", names)
+    }
+}
+
+pub(crate) fn run_status(json_mode: bool) -> Result<()> {
+    let report = crate::agent_hooks_installer::status();
+    if json_mode {
+        println!(
+            "{}",
+            serde_json::to_string_pretty(&report)
+                .unwrap_or_else(|_| serde_json::to_string(&report).unwrap_or_default())
+        );
+    } else {
+        format_status_human(&report);
+    }
+    Ok(())
+}
+
+pub(crate) fn run_uninstall(cli: HooksCliFilter, json_mode: bool) -> Result<()> {
+    let report = crate::agent_hooks_installer::uninstall(cli.into_scope());
+    if json_mode {
+        println!(
+            "{}",
+            serde_json::to_string_pretty(&report)
+                .unwrap_or_else(|_| serde_json::to_string(&report).unwrap_or_default())
+        );
+    } else {
+        format_uninstall_human(&report);
+    }
+    if report.succeeded() {
+        Ok(())
+    } else {
+        anyhow::bail!("one or more hook uninstall steps failed")
+    }
+}
+
+fn format_status_human(r: &crate::agent_hooks_installer::StatusReport) {
+    let path_suffix = r
+        .bundle_source
+        .path
+        .as_deref()
+        .map(|p| format!(" ({})", p))
+        .unwrap_or_default();
+    println!(
+        "{}",
+        t!(
+            "hooks.bundle_source",
+            source = r.bundle_source.kind,
+            path_suffix = path_suffix,
+        )
+    );
+    println!();
+    for c in &r.clis {
+        let summary = if !c.binary_on_path {
+            t!("hooks.cli_not_on_path").into_owned()
+        } else if c.plugin_installed && c.plugin_enabled && c.marketplace_path_valid {
+            t!("hooks.installed").into_owned()
+        } else if c.plugin_installed && !c.marketplace_path_valid {
+            t!("hooks.marketplace_path_stale").into_owned()
+        } else if c.plugin_installed {
+            t!("hooks.installed_but_disabled").into_owned()
+        } else {
+            t!("hooks.not_installed").into_owned()
+        };
+        let detail = format!(
+            "marketplace={}, path_valid={}, plugin={}, enabled={}{}",
+            yn(c.marketplace_registered),
+            yn(c.marketplace_path_valid),
+            yn(c.plugin_installed),
+            yn(c.plugin_enabled),
+            c.detection_fallback
+                .map(|m| format!(", detection={}", m))
+                .unwrap_or_default(),
+        );
+        println!("  {:<10} {:<28}  ({})", c.name, summary, detail);
+        if let Some(p) = c.marketplace_path.as_deref() {
+            println!("    path: {}", p);
+        }
+    }
+}
+
+fn format_uninstall_human(r: &crate::agent_hooks_installer::UninstallReport) {
+    for c in &r.clis {
+        let summary = if !c.attempted {
+            t!("hooks.uninstall_skipped").into_owned()
+        } else {
+            let plugin = c
+                .plugin_uninstalled
+                .map(|b| if b { "ok" } else { "failed" })
+                .unwrap_or("-");
+            let mkt = c
+                .marketplace_removed
+                .map(|b| if b { "ok" } else { "failed" })
+                .unwrap_or("-");
+            format!(
+                "plugin={} marketplace={} staging={}",
+                plugin,
+                mkt,
+                if c.staging_dir_removed {
+                    "ok"
+                } else {
+                    "failed"
+                },
+            )
+        };
+        println!("  {:<10} {}", c.name, summary);
+        for m in &c.messages {
+            println!("    \u{00b7} {}", m);
+        }
+    }
+}
+
+fn yn(b: bool) -> &'static str {
+    if b {
+        "yes"
+    } else {
+        "no"
+    }
+}
