+++ tools/wta/src/cli/mod.rs
@@ -1,1 +1,4 @@
 pub(crate) mod args;
+pub(crate) mod hooks;
+pub(crate) mod probes;
+pub(crate) mod sessions;
