+++ tools/wta/src/main.rs
@@ -46,18 +46,14 @@ mod wsl;
 mod wsl_acp;
 mod wt_protocol_events;
 
-use agent_client_protocol as acp;
 use anyhow::{bail, Context, Result};
 use clap::Parser;
 use serde_json::json;
 use std::sync::Arc;
-use tokio_util::compat::{TokioAsyncReadCompatExt, TokioAsyncWriteCompatExt};
 
-use cli::args::{
-    Cli, Command, HooksAction, HooksCliFilter, InitialView, SessionsAction,
-};
+use cli::args::{Cli, Command, HooksAction, InitialView, SessionsAction};
 #[cfg(test)]
-use cli::args::SessionsOriginArg;
+use cli::args::{HooksCliFilter, SessionsOriginArg};
 use shell::wt_channel::{CliChannel, WtChannel};
 use shell::ShellManager;
 
@@ -464,31 +460,33 @@ async fn main() -> Result<()> {
         // ── Master session registry CLI ──
         Some(Command::Sessions { action }) => match action {
             SessionsAction::List { master, origin } => {
-                run_sessions_list(master, origin.to_filter(), json_mode).await
+                cli::sessions::run_list(master, origin.to_filter(), json_mode).await
             }
         },
 
         // ── Manage agent hooks (install/status/uninstall) ──
         Some(Command::Hooks { action }) => match action {
-            HooksAction::Install { cli } => run_hooks_install(cli),
-            HooksAction::Status => run_hooks_status(json_mode),
-            HooksAction::Uninstall { cli } => run_hooks_uninstall(cli, json_mode),
+            HooksAction::Install { cli } => cli::hooks::run_install(cli),
+            HooksAction::Status => cli::hooks::run_status(json_mode),
+            HooksAction::Uninstall { cli } => cli::hooks::run_uninstall(cli, json_mode),
         },
 
         // ── ACP model list probe ──
-        Some(Command::ProbeModels { agent }) => run_probe_models(&agent).await,
+        Some(Command::ProbeModels { agent }) => cli::probes::run_models(&agent).await,
         Some(Command::ProbeAgentSources { wsl_distro }) => {
-            run_probe_agent_sources(&wsl_distro).await
+            cli::probes::run_agent_sources(&wsl_distro).await
         }
 
         // ── ACP session/list probe (diagnostic) ──
-        Some(Command::ProbeSessions { agent }) => run_probe_sessions(&agent).await,
+        Some(Command::ProbeSessions { agent }) => cli::probes::run_sessions(&agent).await,
 
         // ── Filtered host ACP history probe (diagnostic) ──
-        Some(Command::ProbeHostSessions { agent }) => run_probe_host_sessions(&agent).await,
+        Some(Command::ProbeHostSessions { agent }) => cli::probes::run_host_sessions(&agent).await,
 
         // ── WSL ACP history-scan probe (diagnostic) ──
-        Some(Command::ProbeWslSessions { cli }) => run_probe_wsl_sessions(cli.as_deref()).await,
+        Some(Command::ProbeWslSessions { cli }) => {
+            cli::probes::run_wsl_sessions(cli.as_deref()).await
+        }
 
         // ── No subcommand: a singleton-service mode, or an error. There
         //    is no standalone/default ACP TUI mode — the direct agent-spawn
@@ -561,425 +559,6 @@ fn process_label(cli: &Cli) -> String {
     }
 }
 
-/// Drive [`protocol::acp::probe::probe_models`] on a tokio `LocalSet`
-/// (the ACP client connection is `!Send`), serialize the result to
-/// stdout, force-exit. See exit notes below.
-async fn run_probe_models(agent: &str) -> Result<()> {
-    // Logging is initialized in `main()` (file, not stderr — the Settings UI
-    // captures our stdout for the JSON payload and stderr would pollute it).
-    tracing::info!("probe-models start: agent={}", agent);
-
-    let local = tokio::task::LocalSet::new();
-    let result = match local
-        .run_until(protocol::acp::probe::probe_models(agent))
-        .await
-    {
-        Ok(r) => r,
-        Err(e) => {
-            tracing::error!("probe-models failed: {:#}", e);
-            eprintln!("probe-models failed: {:#}", e);
-            let _ = std::io::Write::flush(&mut std::io::stderr());
-            // Flush the file appender — process::exit skips the guard drop.
-            logging::shutdown_flush();
-            // See exit rationale below.
-            std::process::exit(1);
-        }
-    };
-    tracing::info!(
-        "probe-models ok: {} model(s), current={:?}",
-        result.available_models.len(),
-        result.current_model_id
-    );
-    let payload = serde_json::to_string(&result).context("serialize probe result")?;
-    println!("{}", payload);
-
-    // Force-exit before the tokio runtime tries to drop. The agent we
-    // spawned is e.g. `cmd /c npx ...`; kill_on_drop kills cmd but
-    // the npx → node grandchildren survive as orphans. Tokio's IOCP
-    // reactor stays blocked on handles those orphans inherited and
-    // the runtime drop hangs for ~35s. Runtime cleanup is meaningless
-    // for a one-shot CLI — the caller is blocked on our process
-    // handle, exit now. Orphan grandchildren self-exit shortly after
-    // when they notice their pipes are broken.
-    let _ = std::io::Write::flush(&mut std::io::stdout());
-    // Flush the file appender — process::exit skips the guard drop.
-    logging::shutdown_flush();
-    std::process::exit(0);
-}
-
-#[derive(serde::Serialize)]
-struct AgentSourceProbeEntry {
-    id: &'static str,
-    display_name: &'static str,
-}
-
-#[derive(serde::Serialize)]
-struct AgentSourceProbeResult {
-    wsl_distro: String,
-    agents: Vec<AgentSourceProbeEntry>,
-}
-
-async fn run_probe_agent_sources(wsl_distro: &str) -> Result<()> {
-    let distro = wsl_distro.trim();
-    anyhow::ensure!(!distro.is_empty(), "--wsl-distro must not be empty");
-
-    use futures::StreamExt as _;
-    let agents = futures::stream::iter(crate::agent_registry::KNOWN_AGENTS)
-        .map(|profile| async move {
-            crate::agent_check::wsl_agent_available(distro, profile.id)
-                .await
-                .then_some(AgentSourceProbeEntry {
-                    id: profile.id,
-                    display_name: profile.display_name,
-                })
-        })
-        .buffer_unordered(crate::agent_registry::KNOWN_AGENTS.len())
-        .filter_map(async move |entry| entry)
-        .collect()
-        .await;
-
-    println!(
-        "{}",
-        serde_json::to_string(&AgentSourceProbeResult {
-            wsl_distro: distro.to_string(),
-            agents,
-        })
-        .context("serialize agent source probe")?
-    );
-    Ok(())
-}
-
-/// Drive [`protocol::acp::probe::probe_sessions`] on a tokio `LocalSet`
-/// (the ACP client connection is `!Send`), print the result as pretty
-/// JSON to stdout, force-exit. Diagnostic-only: evaluates whether an
-/// agent CLI answers ACP `session/list` and what it returns.
-async fn run_probe_sessions(agent: &str) -> Result<()> {
-    tracing::info!("probe-sessions start: agent={}", agent);
-
-    let local = tokio::task::LocalSet::new();
-    let result = match local
-        .run_until(protocol::acp::probe::probe_sessions(agent))
-        .await
-    {
-        Ok(r) => r,
-        Err(e) => {
-            tracing::error!("probe-sessions failed: {:#}", e);
-            eprintln!("probe-sessions failed: {:#}", e);
-            let _ = std::io::Write::flush(&mut std::io::stderr());
-            logging::shutdown_flush();
-            std::process::exit(1);
-        }
-    };
-    tracing::info!(
-        "probe-sessions ok: list_ok={} sessions={} err={:?}",
-        result.list_sessions_ok,
-        result.sessions.len(),
-        result.list_sessions_error
-    );
-    let payload = serde_json::to_string_pretty(&result).context("serialize session probe")?;
-    println!("{payload}");
-
-    // Same force-exit rationale as run_probe_models (orphan npx/node
-    // grandchildren keep the tokio reactor blocked on drop).
-    let _ = std::io::Write::flush(&mut std::io::stdout());
-    logging::shutdown_flush();
-    std::process::exit(0);
-}
-
-/// Diagnostic host-history smoke test: run one ACP CLI, fetch
-/// `session/list`, apply the production Class-A filter, and print the
-/// rows in the same compact shape used by the WSL probe.
-async fn run_probe_host_sessions(agent: &str) -> Result<()> {
-    use crate::agent_sessions::{CliSource, SessionLocation};
-    use std::time::Duration;
-
-    tracing::info!("probe-host-sessions start: agent={}", agent);
-
-    // Resolve the CliSource from the agent command so the probe labels and
-    // classifies rows the way production seeding does (which uses the real
-    // `state.cli_source`), instead of assuming Copilot for every agent.
-    let cli_source = CliSource::parse(Some(crate::agent_registry::resolve_agent_id_from_cmd(
-        agent,
-    )));
-
-    let local = tokio::task::LocalSet::new();
-    let rows = match local
-        .run_until(async {
-            let mut spawned = crate::protocol::acp::spawn::spawn_agent_process(agent, None)?;
-            let label = format!("host:{}", crate::session_history::cli_label(&cli_source));
-            let init_timeout = Duration::from_secs(if spawned.is_npx { 25 } else { 10 });
-            let result = crate::protocol::acp::session_list::fetch_session_list(
-                &mut spawned.child,
-                &label,
-                init_timeout,
-                Duration::from_secs(10),
-            )
-            .await;
-            let _ = spawned.child.start_kill();
-            let (_init, list_result) = result?;
-            // session/list unsupported (e.g. `Method not found`) is the production
-            // "empty history, no fallback" case — surface it as `[]` + exit 0, not a
-            // diagnostic failure. A genuine spawn/init error still propagates above.
-            let sessions = list_result.unwrap_or_else(|e| {
-                tracing::info!("probe-host-sessions: session/list unavailable ({e}); returning []");
-                Vec::new()
-            });
-            let idx = crate::agent_pane_origin::load_default_set();
-            Ok::<_, anyhow::Error>(crate::session_history::classify_and_map(
-                &sessions,
-                &idx,
-                SessionLocation::Host,
-                &cli_source,
-            ))
-        })
-        .await
-    {
-        Ok(r) => r,
-        Err(e) => {
-            // Same force-exit rationale as run_probe_sessions: orphan npx/node
-            // grandchildren keep the tokio reactor blocked ~35s on drop.
-            tracing::error!("probe-host-sessions failed: {:#}", e);
-            eprintln!("probe-host-sessions failed: {:#}", e);
-            let _ = std::io::Write::flush(&mut std::io::stderr());
-            logging::shutdown_flush();
-            std::process::exit(1);
-        }
-    };
-
-    let json: Vec<_> = rows
-        .iter()
-        .map(|r| {
-            serde_json::json!({
-                "key": r.key,
-                "cli": format!("{:?}", r.cli_source),
-                "title": r.title,
-                "cwd": r.cwd.to_string_lossy(),
-            })
-        })
-        .collect();
-    println!(
-        "{}",
-        serde_json::to_string_pretty(&json).context("serialize host session probe")?
-    );
-
-    tracing::info!("probe-host-sessions ok: {} row(s)", rows.len());
-    let _ = std::io::Write::flush(&mut std::io::stdout());
-    logging::shutdown_flush();
-    std::process::exit(0);
-}
-
-/// Drive the production WSL ACP history scan
-/// ([`wsl_acp::scan_running_distros_acp`]) on a tokio `LocalSet` (the ACP
-/// connection is `!Send`) and print the discovered sessions as JSON.
-/// Diagnostic-only: exercises the real `wsl.exe` spawn + `session/list`
-/// path that seeds the `/sessions` view.
-async fn run_probe_wsl_sessions(cli: Option<&str>) -> Result<()> {
-    use crate::agent_sessions::CliSource;
-    tracing::info!("probe-wsl-sessions start: cli={:?}", cli);
-
-    let filter: Option<CliSource> = match cli {
-        None => None,
-        Some("copilot") => Some(CliSource::Copilot),
-        Some("claude") => Some(CliSource::Claude),
-        Some("codex") => Some(CliSource::Codex),
-        Some("gemini") => Some(CliSource::Gemini),
-        Some("opencode") => Some(CliSource::OpenCode),
-        Some(other) => {
-            // Reject unknown values rather than silently widening to "scan all"
-            // (Unknown → clis_to_scan → every built-in), which would make the
-            // diagnostic's output contradict the requested restriction.
-            anyhow::bail!(
-                "unknown --cli value {other:?}; expected one of: copilot, claude, codex, gemini, opencode"
-            );
-        }
-    };
-
-    let local = tokio::task::LocalSet::new();
-    let rows = local
-        .run_until(crate::wsl_acp::scan_running_distros_acp(filter.as_ref()))
-        .await;
-
-    let json: Vec<_> = rows
-        .iter()
-        .map(|r| {
-            serde_json::json!({
-                "key": r.key,
-                "cli": format!("{:?}", r.cli_source),
-                "title": r.title,
-                "cwd": r.cwd.to_string_lossy(),
-                "distro": r.location.distro(),
-            })
-        })
-        .collect();
-    println!(
-        "{}",
-        serde_json::to_string_pretty(&json).context("serialize WSL session probe")?
-    );
-
-    tracing::info!("probe-wsl-sessions ok: {} row(s)", rows.len());
-    // Force-exit like the other probes: a distro CLI may leave orphan
-    // grandchildren that keep the tokio reactor blocked on drop.
-    let _ = std::io::Write::flush(&mut std::io::stdout());
-    logging::shutdown_flush();
-    std::process::exit(0);
-}
-
-// ─── Hooks subcommand handlers ──────────────────────────────────────────────
-
-fn run_hooks_install(cli: HooksCliFilter) -> Result<()> {
-    // Logging is initialized in `main()`; the install attempt is observable in
-    // %LOCALAPPDATA%\IntelligentTerminal\logs\wta-install-hooks.log.
-    let scope = cli.into_scope();
-    agent_hooks_installer::ensure_installed_scoped(scope);
-
-    // Verify the install actually landed by checking on-disk status.
-    // ensure_installed_scoped is fire-and-forget (silent on failure),
-    // so we inspect the result independently. `status_scoped(scope)`
-    // skips the Node-CLI spawns for CLIs outside the requested scope —
-    // a `--cli copilot` install no longer pays for `claude plugin list`
-    // and `gemini extensions list` (each ~1-3s of Node startup).
-    let report = agent_hooks_installer::status_scoped(scope);
-    let failed: Vec<&str> = report
-        .clis
-        .iter()
-        .filter(|c| {
-            let in_scope = match scope {
-                agent_hooks_installer::CliScope::All => true,
-                agent_hooks_installer::CliScope::One(kind) => c.name == kind.name(),
-            };
-            // A CLI is "failed" if it's in scope, present on the machine
-            // (cli_found), but hooks are not installed.
-            in_scope && c.binary_on_path && !c.plugin_installed
-        })
-        .map(|c| c.name)
-        .collect();
-
-    if failed.is_empty() {
-        println!("{}", t!("hooks.install_attempted"));
-        Ok(())
-    } else {
-        let names = failed.join(", ");
-        tracing::error!(target: "agent_hooks", clis = %names, "hooks install verification failed");
-        anyhow::bail!("hooks installation failed for: {}", names)
-    }
-}
-
-fn run_hooks_status(json_mode: bool) -> Result<()> {
-    let report = agent_hooks_installer::status();
-    if json_mode {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&report)
-                .unwrap_or_else(|_| serde_json::to_string(&report).unwrap_or_default())
-        );
-    } else {
-        format_hooks_status_human(&report);
-    }
-    Ok(())
-}
-
-fn run_hooks_uninstall(cli: HooksCliFilter, json_mode: bool) -> Result<()> {
-    let report = agent_hooks_installer::uninstall(cli.into_scope());
-    if json_mode {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&report)
-                .unwrap_or_else(|_| serde_json::to_string(&report).unwrap_or_default())
-        );
-    } else {
-        format_hooks_uninstall_human(&report);
-    }
-    if report.succeeded() {
-        Ok(())
-    } else {
-        anyhow::bail!("one or more hook uninstall steps failed")
-    }
-}
-
-fn format_hooks_status_human(r: &agent_hooks_installer::StatusReport) {
-    let path_suffix = r
-        .bundle_source
-        .path
-        .as_deref()
-        .map(|p| format!(" ({})", p))
-        .unwrap_or_default();
-    println!(
-        "{}",
-        t!(
-            "hooks.bundle_source",
-            source = r.bundle_source.kind,
-            path_suffix = path_suffix,
-        )
-    );
-    println!();
-    for c in &r.clis {
-        let summary = if !c.binary_on_path {
-            t!("hooks.cli_not_on_path").into_owned()
-        } else if c.plugin_installed && c.plugin_enabled && c.marketplace_path_valid {
-            t!("hooks.installed").into_owned()
-        } else if c.plugin_installed && !c.marketplace_path_valid {
-            t!("hooks.marketplace_path_stale").into_owned()
-        } else if c.plugin_installed {
-            t!("hooks.installed_but_disabled").into_owned()
-        } else {
-            t!("hooks.not_installed").into_owned()
-        };
-        let detail = format!(
-            "marketplace={}, path_valid={}, plugin={}, enabled={}{}",
-            yn(c.marketplace_registered),
-            yn(c.marketplace_path_valid),
-            yn(c.plugin_installed),
-            yn(c.plugin_enabled),
-            c.detection_fallback
-                .map(|m| format!(", detection={}", m))
-                .unwrap_or_default(),
-        );
-        println!("  {:<10} {:<28}  ({})", c.name, summary, detail);
-        if let Some(p) = c.marketplace_path.as_deref() {
-            println!("    path: {}", p);
-        }
-    }
-}
-
-fn format_hooks_uninstall_human(r: &agent_hooks_installer::UninstallReport) {
-    for c in &r.clis {
-        let summary = if !c.attempted {
-            t!("hooks.uninstall_skipped").into_owned()
-        } else {
-            let plugin = c
-                .plugin_uninstalled
-                .map(|b| if b { "ok" } else { "failed" })
-                .unwrap_or("-");
-            let mkt = c
-                .marketplace_removed
-                .map(|b| if b { "ok" } else { "failed" })
-                .unwrap_or("-");
-            format!(
-                "plugin={} marketplace={} staging={}",
-                plugin,
-                mkt,
-                if c.staging_dir_removed {
-                    "ok"
-                } else {
-                    "failed"
-                },
-            )
-        };
-        println!("  {:<10} {}", c.name, summary);
-        for m in &c.messages {
-            println!("    \u{00b7} {}", m);
-        }
-    }
-}
-
-fn yn(b: bool) -> &'static str {
-    if b {
-        "yes"
-    } else {
-        "no"
-    }
-}
-
 // ─── Helper: connect to WT COM protocol (no debug channel, no ShellManager) ─────────
 
 async fn connect_channel() -> Result<CliChannel> {
@@ -1039,324 +618,6 @@ async fn get_first_tab_id(channel: &CliChannel, window_id: &str) -> Result<Strin
             _ => None,
         })
         .ok_or_else(|| anyhow::anyhow!("{}", t!("output.no_tabs_in_window", window_id = window_id)))
-}
-
-// ─── sessions CLI helpers ───────────────────────────────────────────────────
-
-const MASTER_NOT_RUNNING: &str = "wta-master not running. Start Windows Terminal first.";
-
-async fn run_sessions_list(
-    master_override: Option<String>,
-    origin_filter: agent_sessions::OriginFilter,
-    json_mode: bool,
-) -> Result<()> {
-    let local = tokio::task::LocalSet::new();
-    let sessions = local
-        .run_until(fetch_sessions_from_master(master_override))
-        .await?;
-    // Origin filter is applied client-side: master always returns the
-    // full registry so this command can act as the debug eye-of-god
-    // view (default `--origin all`). `--origin shell` matches what
-    // the MVP sessions picker shows; `--origin agent-pane` surfaces the
-    // rows MVP sessions hides.
-    let mut filtered: Vec<session_registry::SessionInfo> = sessions
-        .into_iter()
-        .filter(|s| origin_filter.matches_opt(s.origin.as_ref()))
-        .collect();
-    // Match the `/sessions` picker, which renders newest-activity-first.
-    // `None` (no timestamp) sorts last.
-    filtered.sort_by(|a, b| b.last_activity_at_ms.cmp(&a.last_activity_at_ms));
-    if json_mode {
-        print!("{}", format_sessions_json_lines(&filtered)?);
-    } else {
-        print!("{}", format_sessions_table(&filtered));
-    }
-    Ok(())
-}
-
-async fn fetch_sessions_from_master(
-    master_override: Option<String>,
-) -> Result<Vec<session_registry::SessionInfo>> {
-    let pipe_name = resolve_master_pipe(master_override).await?;
-    let pipe = open_master_pipe_for_cli(&pipe_name).await?;
-    let (read_half, write_half) = tokio::io::split(pipe);
-    let outgoing = write_half.compat_write();
-    let incoming = read_half.compat();
-    let (conn, handle_io) = crate::protocol::acp::conn::spawn_client(
-        acp::Client.builder().name("wta-sessions"),
-        crate::protocol::acp::conn::byte_streams(outgoing, incoming),
-    );
-    tokio::task::spawn_local(async move {
-        let _ = handle_io.await;
-    });
-
-    let init_started = std::time::Instant::now();
-    let init_result = conn
-        .initialize(
-            acp::schema::v1::InitializeRequest::new(acp::schema::ProtocolVersion::V1)
-                .client_capabilities(acp::schema::v1::ClientCapabilities::new())
-                .client_info(
-                    acp::schema::v1::Implementation::new("wta-sessions", env!("CARGO_PKG_VERSION"))
-                        .title("Windows Terminal Agent sessions CLI"),
-                ),
-        )
-        .await;
-    telemetry::log_acp_initialize_complete(
-        init_started.elapsed().as_secs_f64() * 1000.0,
-        init_result.is_ok(),
-        "SessionsCli",
-        if init_result.is_ok() { "" } else { "AcpError" },
-        init_result
-            .as_ref()
-            .err()
-            .map(|e| e.code.into())
-            .unwrap_or(0),
-    );
-    init_result.map_err(|_| anyhow::anyhow!(MASTER_NOT_RUNNING))?;
-
-    let req = session_registry::build_sessions_list_request(false);
-    let resp = conn
-        .ext_method(req)
-        .await
-        .map_err(|_| anyhow::anyhow!(MASTER_NOT_RUNNING))?;
-    let parsed = session_registry::parse_sessions_list_response(&resp.0)
-        .context("parse sessions/list response")?;
-    Ok(parsed.sessions)
-}
-
-/// Best-effort: register a WTA-launched CLI session with `wta-master` as a
-/// *born-bound* row — bound to its pane, with no hooks involved. Sends a
-/// `SessionStarted` over the `intellterm.wta/session_born_bound` method, which
-/// the master turns into a Class-B (`origin = Unknown`) row whose
-/// `pane_session_id` is the pane we just created and records as binding-only
-/// (so the file watcher may still supply activity/status when no hook is
-/// installed). Best-effort: if master is unreachable there is no registry to
-/// populate, so the registration is dropped (logged at `warn`) and the tab
-/// still opens normally.
-async fn register_launched_session_with_master(
-    session_id: &str,
-    pane_session_id: &str,
-    cli_id: &str,
-    cwd: Option<&str>,
-    wsl_distro: Option<&str>,
-) {
-    let event = crate::agent_sessions::SessionEvent::SessionStarted {
-        key: session_id.to_string(),
-        cli_source: crate::agent_sessions::CliSource::from(
-            crate::session_registry::SessionHookCliSource::Known(cli_id.to_string()),
-        ),
-        pane_session_id: pane_session_id.to_string(),
-        cwd: cwd.map(std::path::PathBuf::from).unwrap_or_default(),
-        // Empty title: the master refreshes the row's title from the CLI's
-        // on-disk session artefacts once they appear.
-        title: String::new(),
-    };
-    // A WSL delegate carries its distro so the master stamps the row
-    // `Wsl { distro }` → the session view shows the `[WSL-<distro>]` prefix.
-    let req = match wsl_distro {
-        Some(distro) => session_registry::build_born_bound_request_wsl(&event, distro),
-        None => session_registry::build_born_bound_request(&event),
-    };
-
-    // Own LocalSet so the `spawn_local` transport works regardless of how the
-    // delegate's runtime was set up (mirrors `run_sessions_list`).
-    let local = tokio::task::LocalSet::new();
-    let result: Result<()> = local
-        .run_until(async move {
-            let pipe_name = resolve_master_pipe(None).await?;
-            let pipe = open_master_pipe_for_cli(&pipe_name).await?;
-            let (read_half, write_half) = tokio::io::split(pipe);
-            let outgoing = write_half.compat_write();
-            let incoming = read_half.compat();
-            let (conn, handle_io) = crate::protocol::acp::conn::spawn_client(
-                acp::Client.builder().name("wta-delegate"),
-                crate::protocol::acp::conn::byte_streams(outgoing, incoming),
-            );
-            tokio::task::spawn_local(async move {
-                let _ = handle_io.await;
-            });
-
-            conn.initialize(
-                acp::schema::v1::InitializeRequest::new(acp::schema::ProtocolVersion::V1)
-                    .client_capabilities(acp::schema::v1::ClientCapabilities::new())
-                    .client_info(
-                        acp::schema::v1::Implementation::new(
-                            "wta-delegate",
-                            env!("CARGO_PKG_VERSION"),
-                        )
-                        .title("Windows Terminal Agent delegate"),
-                    ),
-            )
-            .await
-            .map_err(|_| anyhow::anyhow!(MASTER_NOT_RUNNING))?;
-
-            conn.ext_method(req)
-                .await
-                .map_err(|_| anyhow::anyhow!(MASTER_NOT_RUNNING))?;
-            Ok(())
-        })
-        .await;
-
-    if let Err(e) = result {
-        tracing::warn!(
-            target: "delegate",
-            error = %e,
-            "register born-bound session with master failed (best-effort)"
-        );
-    }
-}
-
-async fn resolve_master_pipe(master_override: Option<String>) -> Result<String> {
-    if let Some(pipe) = master_override.filter(|s| !s.trim().is_empty()) {
-        return Ok(pipe);
-    }
-
-    for attempt in 0..2 {
-        if let Some(path) = runtime_paths::master_pipe_file_path() {
-            if let Ok(contents) = std::fs::read_to_string(path) {
-                let pipe = contents.trim();
-                if !pipe.is_empty() {
-                    return Ok(pipe.to_string());
-                }
-            }
-        }
-        if attempt == 0 {
-            tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-        }
-    }
-    Err(anyhow::anyhow!(MASTER_NOT_RUNNING))
-}
-
-async fn open_master_pipe_for_cli(
-    pipe_name: &str,
-) -> Result<tokio::net::windows::named_pipe::NamedPipeClient> {
-    for attempt in 0..2 {
-        match tokio::net::windows::named_pipe::ClientOptions::new().open(pipe_name) {
-            Ok(pipe) => return Ok(pipe),
-            Err(_) if attempt == 0 => {
-                tokio::time::sleep(std::time::Duration::from_millis(100)).await
-            }
-            Err(_) => return Err(anyhow::anyhow!(MASTER_NOT_RUNNING)),
-        }
-    }
-    Err(anyhow::anyhow!(MASTER_NOT_RUNNING))
-}
-
-fn format_sessions_json_lines(sessions: &[session_registry::SessionInfo]) -> Result<String> {
-    let mut out = String::new();
-    for session in sessions {
-        out.push_str(&serde_json::to_string(session)?);
-        out.push('\n');
-    }
-    Ok(out)
-}
-
-fn format_sessions_table(sessions: &[session_registry::SessionInfo]) -> String {
-    let mut out = String::new();
-    if sessions.is_empty() {
-        out.push_str("No sessions.\n");
-        return out;
-    }
-    out.push_str(&format!(
-        "{:<4} {:<24} {:<10} {:<10} {:<10} {:<16} {:<20} {:<20} {}\n",
-        "#", "SESSION", "STATUS", "CLI", "ORIGIN", "LOCATION", "PANE", "UPDATED", "TITLE"
-    ));
-    for (i, session) in sessions.iter().enumerate() {
-        let sid = session.session_id.to_string();
-        let short_sid = if sid.len() > 24 {
-            &sid[..24]
-        } else {
-            sid.as_str()
-        };
-        out.push_str(&format!(
-            "{:<4} {:<24} {:<10} {:<10} {:<10} {:<16} {:<20} {:<20} {}\n",
-            i + 1,
-            short_sid,
-            status_label(session.status.as_ref()),
-            cli_source_label(session.cli_source.as_ref()),
-            origin_label(session.origin.as_ref()),
-            location_label(&session.location),
-            session.pane_session_id.as_deref().unwrap_or("-"),
-            updated_label(session),
-            session.title.as_deref().unwrap_or("-"),
-        ));
-    }
-    out
-}
-
-fn status_label(status: Option<&agent_sessions::AgentStatus>) -> String {
-    status
-        .map(|s| format!("{s:?}"))
-        .unwrap_or_else(|| "-".to_string())
-}
-
-fn cli_source_label(source: Option<&agent_sessions::CliSource>) -> String {
-    match source {
-        Some(agent_sessions::CliSource::Claude) => "Claude".to_string(),
-        Some(agent_sessions::CliSource::Codex) => "Codex".to_string(),
-        Some(agent_sessions::CliSource::Copilot) => "Copilot".to_string(),
-        Some(agent_sessions::CliSource::Gemini) => "Gemini".to_string(),
-        Some(agent_sessions::CliSource::OpenCode) => "OpenCode".to_string(),
-        Some(agent_sessions::CliSource::Unknown(s)) if !s.is_empty() => s.clone(),
-        _ => "-".to_string(),
-    }
-}
-
-/// Render a `SessionOrigin` for the `wta sessions list` table. `None`
-/// is the on-the-wire representation for "field absent" (legacy rows
-/// or notification paths that don't carry origin) — we print `-`
-/// rather than fabricating an origin so the operator can tell
-/// "untagged" from "shell".
-fn origin_label(origin: Option<&agent_sessions::SessionOrigin>) -> &'static str {
-    match origin {
-        Some(agent_sessions::SessionOrigin::AgentPane) => "AgentPane",
-        Some(agent_sessions::SessionOrigin::Unknown) => "Shell",
-        None => "-",
-    }
-}
-
-/// Render a `SessionLocation` for the `wta sessions list` table: `host`
-/// for Windows-profile sessions, `wsl:<distro>` for sessions discovered
-/// inside a WSL distro.
-fn location_label(location: &agent_sessions::SessionLocation) -> String {
-    match location {
-        agent_sessions::SessionLocation::Host => "host".to_string(),
-        agent_sessions::SessionLocation::Wsl { distro } => format!("wsl:{distro}"),
-    }
-}
-
-/// Render the UPDATED column. Prefers the `updated_at` ISO string (set for
-/// live sessions); for history-scanned rows that only carry an epoch-ms
-/// `last_activity_at_ms`, formats that as a `YYYY-MM-DD HH:MM` UTC stamp so
-/// the column isn't blank. `-` when neither is available.
-fn updated_label(s: &session_registry::SessionInfo) -> String {
-    if let Some(u) = s.updated_at.as_deref() {
-        return u.to_string();
-    }
-    match s.last_activity_at_ms {
-        Some(ms) => format_epoch_ms_utc(ms),
-        None => "-".to_string(),
-    }
-}
-
-/// Format epoch milliseconds as `YYYY-MM-DD HH:MM` (UTC) without pulling in a
-/// date crate. Uses Howard Hinnant's `civil_from_days` algorithm.
-fn format_epoch_ms_utc(ms: u64) -> String {
-    let secs = (ms / 1000) as i64;
-    let days = secs.div_euclid(86_400);
-    let tod = secs.rem_euclid(86_400);
-    let (hour, min) = (tod / 3600, (tod % 3600) / 60);
-    // civil_from_days: days since 1970-01-01 -> (year, month, day).
-    let z = days + 719_468;
-    let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = z - era * 146_097;
-    let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365;
-    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
-    let mp = (5 * doy + 2) / 153;
-    let day = doy - (153 * mp + 2) / 5 + 1;
-    let month = if mp < 10 { mp + 3 } else { mp - 9 };
-    let year = yoe + era * 400 + if month <= 2 { 1 } else { 0 };
-    format!("{year:04}-{month:02}-{day:02} {hour:02}:{min:02}")
 }
 
 // ─── Output helpers ─────────────────────────────────────────────────────────
@@ -1988,7 +1249,7 @@ async fn delegate_with_context(
                 if let (Some(sid), Some(pane)) =
                     (pinned_session_id.as_deref(), pane_guid.as_deref())
                 {
-                    register_launched_session_with_master(
+                    cli::sessions::register_launched(
                         sid,
                         pane,
                         &runtime.id,
@@ -2033,7 +1294,7 @@ async fn delegate_with_context(
     // bind them with no hooks (best-effort). Only when both are known —
     // i.e. a pinnable agent (Copilot/Claude/Gemini) whose tab was created.
     if let (Some(sid), Some(pane)) = (pinned_session_id.as_deref(), pane_guid.as_deref()) {
-        register_launched_session_with_master(sid, pane, &runtime.id, cwd, None).await;
+        cli::sessions::register_launched(sid, pane, &runtime.id, cwd, None).await;
     }
 
     Ok(())
